// Code/src/app/shared/model/LoginModel.ts
import { PackingModel } from './PackingModel';
export class LoginModel {
    Id: number;
    FirstName: string;
    LastName: string;
    Address: string;
    Gender: string;
    UserType: string;
    EmailId: string;
    Mobile: string;
    UserName: string;
    Password: string;
    DOC: any;
    DOM: any;
    IsActive: any;
}